Fixed support for non-dryml-based, erb-based gems like ckeditor
Needed to allow for modules which did not adhere to dryml standards.
Specifically there was one place where it was assumed a module would be
found when searched for but sometimes was not found. This required a
check for a null module before "processing it". In a second place, it
was assumed that a certain method would be pre-defined (namely the
#name method) but it was not. In this case we needed a begin-rescue-end
construct since it seemed that even though a "name" method was not
defined, some cases expected this scenario and worked around it via a
"method not defined" treatment. Others, as in the case of CKEditor's
gem, just simply failed. The begin-rescue-end allowed the call to
attempt to go through and, if it was not caught by a no method defined
block, the rescue block allowed processing to continue. I am not sure
if any specific action should be taken within the rescue block so I
left it blank. In the case of the CKEditor gem integration testing I
did, this empty rescue block seemed to work fiine but the general case
still needs further examination.